Move contants file to separeate data directory
