--- src/contants.js
@@ -1,13 +1,0 @@
-import htmlSvg from "./images/skills/html.svg";
-import javascriptSvg from "./images/skills/javascript.svg";
-import reactSvg from "./images/skills/react.svg";
-import tailwindSvg from "./images/skills/tailwind.svg";
-
-export const navLinks = ["home", "about", "projects", "contact"];
-
-export const socialLinks = {
-  linkedin: "https://www.linkedin.com/in/joviandsouza",
-  github: "https://github.com/Jovian-Dsouza",
-};
-
-export const techStacks = [htmlSvg, javascriptSvg, reactSvg, tailwindSvg];
